double->float

+++ main.cpp
@@ -17,9 +17,9 @@
 
 struct vec3
 	{
-	double x;
-	double y;
-	double z;
+	float x;
+	float y;
+	float z;
 	};
 
 struct framestamp
@@ -33,30 +33,30 @@ struct dataset
 	std::vector<framestamp> frames;
 	};
 
-local std::vector<vec3> rotate(std::vector<vec3> input, double yaw, double pitch, double roll)
+local std::vector<vec3> rotate(std::vector<vec3> input, float yaw, float pitch, float roll)
 {
-	const double ca = cos(yaw);
-	const double cb = cos(pitch);
-	const double cy = cos(roll);
-	const double sa = sin(yaw);
-	const double sb = sin(pitch);
-	const double sy = sin(roll);
+	const float ca = cos(yaw);
+	const float cb = cos(pitch);
+	const float cy = cos(roll);
+	const float sa = sin(yaw);
+	const float sb = sin(pitch);
+	const float sy = sin(roll);
 
-	const double m00 = ca * cb;
-	const double m01 = ca * sb * sy - sa * cy;
-	const double m02 = ca * sb * cy + sa * sy;
-	const double m10 = sa * cb;
-	const double m11 = sa * sb * sy + ca * cy;
-	const double m12 = sa * sb * cy - ca * sy;
-	const double m20 = -sb;
-	const double m21 = cb * sy;
-	const double m22 = cb * cy;
+	const float m00 = ca * cb;
+	const float m01 = ca * sb * sy - sa * cy;
+	const float m02 = ca * sb * cy + sa * sy;
+	const float m10 = sa * cb;
+	const float m11 = sa * sb * sy + ca * cy;
+	const float m12 = sa * sb * cy - ca * sy;
+	const float m20 = -sb;
+	const float m21 = cb * sy;
+	const float m22 = cb * cy;
 
 	for( unsigned i = 0; i < input.size(); ++i)
 	{
-		const double x = input[i].x;
-		const double y = input[i].y;
-		const double z = input[i].z;
+		const float x = input[i].x;
+		const float y = input[i].y;
+		const float z = input[i].z;
 
 		input[i].x = m00 * x + m01 * y + m02 * z;
 		input[i].y = m10 * x + m11 * y + m12 * z;
@@ -68,10 +68,10 @@ local std::vector<vec3> rotate(std::vector<vec3> input, double yaw, double pitch
 
 local uint64_t randns(uint64_t range)
 {
-	return (double)range * (double)rand() / (double)RAND_MAX;
+	return (float)range * (float)rand() / (float)RAND_MAX;
 }
 
-local vec3 randVec(double mag)
+local vec3 randVec(float mag)
 {
 	vec3 v;
 
@@ -79,7 +79,7 @@ local vec3 randVec(double mag)
 	v.y = rand() - RAND_MAX / 2;
 	v.z = rand() - RAND_MAX / 2;
 
-	double normalize = mag/fsqrt(v.x * v.x + v.y * v.y + v.z * v.z);
+	float normalize = mag/fsqrt(v.x * v.x + v.y * v.y + v.z * v.z);
 
 	v.x *= normalize;
 	v.y *= normalize;
@@ -88,16 +88,16 @@ local vec3 randVec(double mag)
 	return v;
 }
 
-local framestamp filterVisible(framestamp input, double visibleRadiusId, double visibleRadiusOd)
+local framestamp filterVisible(framestamp input, float visibleRadiusId, float visibleRadiusOd)
 {
-	const double vrid2 = visibleRadiusId * visibleRadiusId;
-	const double vrod2 = visibleRadiusOd * visibleRadiusOd;
+	const float vrid2 = visibleRadiusId * visibleRadiusId;
+	const float vrod2 = visibleRadiusOd * visibleRadiusOd;
 	framestamp f = {input.ns};
 
 	for( unsigned i = 0; i < input.points.size(); ++i)
 		{
 		vec3 p = input.points[i];
-		double r = p.x * p.x + p.y * p.y;
+		float r = p.x * p.x + p.y * p.y;
 		//point must not be inside ID, must be inside OD, and must be Z+
 		if(r > vrid2 && r < vrod2 && p.z > 0)
 			{
@@ -107,7 +107,7 @@ local framestamp filterVisible(framestamp input, double visibleRadiusId, double 
 	return f;
 }
 
-local dataset genData(unsigned n, double d_yaw, double d_pitch, double d_roll)
+local dataset genData(unsigned n, float d_yaw, float d_pitch, float d_roll)
 {
 	dataset ret;
 
@@ -121,9 +121,9 @@ local dataset genData(unsigned n, double d_yaw, double d_pitch, double d_roll)
 
 	for( unsigned i = 0; i < n; ++i )
 		{
-		double yaw = d_yaw * ts / 1e9;
-		double pitch = d_pitch * ts / 1e9;
-		double roll = d_roll * ts / 1e9;
+		float yaw = d_yaw * ts / 1e9;
+		float pitch = d_pitch * ts / 1e9;
+		float roll = d_roll * ts / 1e9;
 
 		framestamp frame = {ts};
 		frame.points = rotate(frame0.points, yaw, pitch, roll);
@@ -137,14 +137,14 @@ local dataset genData(unsigned n, double d_yaw, double d_pitch, double d_roll)
 	return ret;
 }
 
-local double dot(vec3 a, vec3 b)
+local float dot(vec3 a, vec3 b)
 {
 return a.x * b.x + a.y * b.y + a.z * b.z;
 }
 
-local uint64_t evalScore(const dataset & data, double d_yaw, double d_pitch, double d_roll)
+local uint64_t evalScore(const dataset & data, float d_yaw, float d_pitch, float d_roll)
 {
-	double accum = 0;
+	float accum = 0;
 	for( unsigned i = 0; i < data.frames.size() - 1; ++i)
 		{
 		const auto & f1 = data.frames[i];
@@ -152,18 +152,18 @@ local uint64_t evalScore(const dataset & data, double d_yaw, double d_pitch, dou
 			{
 			const auto & f2 = data.frames[j];
 			const uint64_t ts = f2.ns - f1.ns;
-			const double yaw = d_yaw * ts / 1e9;
-			const double pitch = d_pitch * ts / 1e9;
-			const double roll = d_roll * ts / 1e9;
+			const float yaw = d_yaw * ts / 1e9;
+			const float pitch = d_pitch * ts / 1e9;
+			const float roll = d_roll * ts / 1e9;
 
 			std::vector<vec3> f1_rotated = rotate(f1.points, yaw, pitch, roll);
-			static const double DOT_LIMIT = .999995;
-//			static const double DOT_LIMIT = .9995;
+			static const float DOT_LIMIT = .999995;
+//			static const float DOT_LIMIT = .9995;
 			for( unsigned k = 0; k < f1_rotated.size(); ++k)
 				{
 				for( unsigned l = 0; l < f2.points.size(); ++l)
 					{
-					double d = dot(f1_rotated[k], f2.points[l]);
+					float d = dot(f1_rotated[k], f2.points[l]);
 					if( d > DOT_LIMIT)
 						{
 						accum += (d - DOT_LIMIT) / (1 - DOT_LIMIT);
@@ -178,9 +178,9 @@ local uint64_t evalScore(const dataset & data, double d_yaw, double d_pitch, dou
 
 int main()
 {
-double d_yaw = RPM_YAW * 2 * M_PI / 60;
-double d_pitch = RPM_PITCH * 2 * M_PI / 60;
-double d_roll = RPM_ROLL * 2 * M_PI / 60;
+float d_yaw = RPM_YAW * 2 * M_PI / 60;
+float d_pitch = RPM_PITCH * 2 * M_PI / 60;
+float d_roll = RPM_ROLL * 2 * M_PI / 60;
 
 std::cout << "Sim Params:"<<d_yaw<<','<<d_pitch<<','<<d_roll << std::endl;
 
@@ -209,7 +209,7 @@ std::cout << "d_pitch,score" << std::endl;
 for( int i = 900; i <= 1100; i += 1)
 //for( int i = -5000; i <= 5000; i += 25)
 	{
-	double dp = d_pitch * (double)i / (double)1000;
+	float dp = d_pitch * (float)i / (float)1000;
 	uint64_t score = evalScore(data, d_yaw, dp, d_roll);
 	std::cout << dp <<','<< score << std::endl;
 	}
